requirements.txt
streamlit
duckdb
requests
pandas
plotly
python-dotenv
schedule
